dom/tv/test/mochitest/test_tv_set_invalid_current_channel.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test an error case for SetCurrentChannel for TV API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript" src="./head.js"></script>
<script type="application/javascript">

function runTest() {
  ok('tv' in navigator, "navigator.tv should exist.");

  navigator.tv.getTuners().then(
    function(aTuners) {
      ok(aTuners.length > 0, "Got at least 1 tuner.");
      var tuner = aTuners[0];
      var selectedSourceType = tuner.getSupportedSourceTypes()[0];

      aTuners[0].getSources().then(
        function(aSources) {
          ok(aSources.length > 0, "Got at least 1 source.");

          aSources[0].setCurrentChannel("NonExistentChannelNumber").then(
            function() {
              ok(false, "Setting an invalid current channel should get error.");
              SimpleTest.finish();
            },
            function(aError) {
              is(aError.name, "AbortError", "AbortError should be expected.");
              SimpleTest.finish();
            }
          );
        },
        function(aError) {
          ok(false, "Error occurred when getting sources: " + aError);
          SimpleTest.finish();
        }
      );
    },
    function(aError) {
      ok(false, "Error occurred when getting tuners: " + aError);
      SimpleTest.finish();
    }
  );
}

SimpleTest.waitForExplicitFinish();
prepareTest(runTest);

</script>
</pre>
</body>
</html>
